moved main file into views folder for displaying home page
